+++ accommodations/tests/test_overview_tab.py
@@ -314,7 +314,11 @@ class AccommodationOverviewTestCase(
             )
         )
 
-        self.assertContains(response, "Change", html=True)
+        self.assertContains(
+            response,
+            'Change<span class="govuk-visually-hidden">record details</span>',
+            html=True,
+        )
 
         response = self.client.get(
             reverse(
@@ -323,7 +327,11 @@ class AccommodationOverviewTestCase(
             )
         )
 
-        self.assertNotContains(response, "Change", html=True)
+        self.assertNotContains(
+            response,
+            'Change<span class="govuk-visually-hidden">record details</span>',
+            html=True,
+        )
 
     def test_duplicate_label_renders_for_duplicate_guests_only(
         self,