+++ browser_tests/pages/home_page.py
@@ -1,4 +1,9 @@
 from .share_page import SharePage
 
 
-class HomePage(SharePage): ...
+class HomePage(SharePage):
+    # TODO: Replace when we have the card updated HFURB-3979
+    def click_on_card(self, card_text):
+        self.page.get_by_role("heading", level=3, name=card_text).locator(
+            "xpath=../../../../a"
+        ).click()