+++ deduplication/views.py
@@ -6,7 +6,6 @@ from typing import Any, List
 
 from crispy_forms_gds.helper import FormHelper
 from crispy_forms_gds.layout import Field, Fieldset, Layout, Size
-from django.conf import settings
 from django.contrib import messages
 from django.db.models import Count, Q
 from django.forms import CheckboxInput, CheckboxSelectMultiple
@@ -32,7 +31,6 @@ from django_tables2 import (
 )
 from formtools.wizard.views import NamedUrlSessionWizardView
 
-from accounts.enums import GroupType
 from case_management.settings import sentry_sdk
 from deduplication.constants import DEDUP_SPONSORS_SEARCH_FIELDS
 from deduplication.forms import (
@@ -65,6 +63,7 @@ from visa_applications.templatetags.visa_application_extras import (
 from webapp.constants import (
     ACCOMMODATION_SEARCH_FIELDS,
     FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES,
+    FIX_GUEST_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES,
     GUEST_SEARCH_FIELDS,
     visa_status_list,
 )
@@ -190,17 +189,16 @@ UNDO_DEDUPLICATION_FORM_TEMPLATES = {
 class SelectRecordTypeView(PermissionsMixin, FormView):
     template_name = "select_duplicate_record_type.html"
     form_class = SelectRecordTypeForm
-
-    def get_group_type(self):
-        if settings.FIX_DUPLICATE_RECORDS_ENABLED:
-            return list(FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
-        return [GroupType.DEV]
+    group_type = list(FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
 
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
-        # TEMP: Only show the 'Guests' option to users in the DEV group
+        # TEMP: Only show the 'Guests' option to users in the DEV or
+        # LOCAL_AUTHORITY_EARLY_ADOPTERS group
         show_guests_option = False
-        if self.request.user.groups.filter(name="dev").exists():
+        if self.request.user.groups.filter(
+            Q(name="dev") | Q(name="local_authority_early_adopters")
+        ).exists():
             show_guests_option = True
         kwargs["show_guests_option"] = show_guests_option
         return kwargs
@@ -406,7 +404,7 @@ class ManualSponsorDeduplicationTable(dj_tables.Table):
     def render_is_eoi(self, value):
         return "True" if value is True else "False"
 
-    def render_select(self, value):
+    def render_select(self, value, record):
         hidden_sponsor_inputs = mark_safe(
             "".join(
                 [
@@ -426,12 +424,14 @@ class ManualSponsorDeduplicationTable(dj_tables.Table):
             '<input type="hidden" name="csrfmiddlewaretoken" value="{csrf_token}"/>'
             '<input type="hidden" name="select-record-sponsor_record"'
             'id="id_select-record-sponsor_record" value="{value}"/>'
-            '<button type="submit" name="submit" class="govuk-link">Select</button>'
+            '<button type="submit" name="submit" class="govuk-link">Select'
+            '<span class="govuk-visually-hidden"> {record_name}</span></button>'
             "</form>",
             management_form=self.context["wizard"]["management_form"],
             value=value,
             csrf_token=get_token(self.request),
             hidden_sponsor_inputs=hidden_sponsor_inputs,
+            record_name=record.get_full_name(),
         )
 
     class Meta:
@@ -697,7 +697,7 @@ class ManualGuestDeduplicationTable(dj_tables.Table):
             value=value,
         )
 
-    def render_select(self, value):
+    def render_select(self, value, record):
         hidden_guest_inputs = mark_safe(
             "".join(
                 [
@@ -717,12 +717,14 @@ class ManualGuestDeduplicationTable(dj_tables.Table):
             '<input type="hidden" name="csrfmiddlewaretoken" value="{csrf_token}"/>'
             '<input type="hidden" name="select-record-guest_record"'
             'id="id_select-record-guest_record" value="{value}"/>'
-            '<button type="submit" name="submit" class="govuk-link">Select</button>'
+            '<button type="submit" name="submit" class="govuk-link">Select'
+            '<span class="govuk-visually-hidden"> {record_name}</span></button>'
             "</form>",
             management_form=self.context["wizard"]["management_form"],
             value=value,
             csrf_token=get_token(self.request),
             hidden_guest_inputs=hidden_guest_inputs,
+            record_name=record.get_full_name(),
         )
 
     def render_visa_status(self, value):
@@ -871,7 +873,7 @@ class ManualAccommodationDeduplicationTable(dj_tables.Table, TableRendererMixin)
             value,
         )
 
-    def render_select(self, value):
+    def render_select(self, value, record):
         hidden_accommodation_inputs = mark_safe(
             "".join(
                 [
@@ -892,12 +894,14 @@ class ManualAccommodationDeduplicationTable(dj_tables.Table, TableRendererMixin)
             '<input type="hidden" name="csrfmiddlewaretoken" value="{csrf_token}"/>'
             '<input type="hidden" name="select-record-accommodation_record"'
             'id="id_select-record-accommodation_record" value="{value}"/>'
-            '<button type="submit" name="submit" class="govuk-link">Select</button>'
+            '<button type="submit" name="submit" class="govuk-link">Select'
+            '<span class="govuk-visually-hidden"> {record_name}</span></button>'
             "</form>",
             management_form=self.context["wizard"]["management_form"],
             value=value,
             csrf_token=get_token(self.request),
             hidden_accommodation_inputs=hidden_accommodation_inputs,
+            record_name=record.full_address,
         )
 
     class Meta:
@@ -1000,7 +1004,7 @@ class ManualViewSelectedSponsorsTable(dj_tables.Table):
     def render_is_eoi(self, value):
         return "True" if value is True else "False"
 
-    def render_remove(self, value):
+    def render_remove(self, value, record):
         return format_html(
             '<form method="post">'
             "{management_form}"
@@ -1008,11 +1012,13 @@ class ManualViewSelectedSponsorsTable(dj_tables.Table):
             '<input type="hidden" '
             'name="review-selected-records-sponsor_record_to_remove"'
             'id="id_review-selected-records-sponsor_record_to_remove" value="{value}"/>'
-            '<button type="submit" name="submit" class="govuk-link">Remove</button>'
+            '<button type="submit" name="submit" class="govuk-link">Remove'
+            '<span class="govuk-visually-hidden"> {record_name}</span></button>'
             "</form>",
             management_form=self.context["wizard"]["management_form"],
             value=value,
             csrf_token=get_token(self.request),
+            record_name=record.get_full_name(),
         )
 
     class Meta:
@@ -1104,18 +1110,20 @@ class ManualViewSelectedGuestsTable(dj_tables.Table):
             value=value,
         )
 
-    def render_remove(self, value):
+    def render_remove(self, value, record):
         return format_html(
             '<form method="post">'
             "{management_form}"
             '<input type="hidden" name="csrfmiddlewaretoken" value="{csrf_token}"/>'
             '<input type="hidden" name="review-selected-records-guest_record_to_remove"'
             'id="id_review-selected-records-guest_record_to_remove" value="{value}"/>'
-            '<button type="submit" name="submit" class="govuk-link">Remove</button>'
+            '<button type="submit" name="submit" class="govuk-link">Remove'
+            '<span class="govuk-visually-hidden"> {record_name}</span></button>'
             "</form>",
             management_form=self.context["wizard"]["management_form"],
             value=value,
             csrf_token=get_token(self.request),
+            record_name=record.get_full_name(),
         )
 
     def render_visa_status(self, value):
@@ -1207,7 +1215,7 @@ class ManualViewSelectedAccommodationsTable(dj_tables.Table, TableRendererMixin)
             value,
         )
 
-    def render_remove(self, value):
+    def render_remove(self, value, record):
         return format_html(
             '<form method="post">'
             "{management_form}"
@@ -1216,11 +1224,13 @@ class ManualViewSelectedAccommodationsTable(dj_tables.Table, TableRendererMixin)
             'name="review-selected-records-accommodation_record_to_remove"'
             'id="id_review-selected-records-accommodation_record_to_remove" '
             'value="{value}"/>'
-            '<button type="submit" name="submit" class="govuk-link">Remove</button>'
+            '<button type="submit" name="submit" class="govuk-link">Remove'
+            '<span class="govuk-visually-hidden"> {record_name}</span></button>'
             "</form>",
             management_form=self.context["wizard"]["management_form"],
             value=value,
             csrf_token=get_token(self.request),
+            record_name=record.full_address,
         )
 
     class Meta:
@@ -2383,11 +2393,7 @@ class UndoDeduplicationRecordsFormWizard(
 class SelectAndReviewSponsorRecordsFormWizard(SelectAndViewRecordsFormWizard):
     model = MvVolunteer
     record_type = "sponsor and host"
-
-    def get_group_type(self):
-        if settings.FIX_DUPLICATE_RECORDS_ENABLED:
-            return list(FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
-        return [GroupType.DEV]
+    group_type = list(FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -2783,7 +2789,7 @@ class UndoDeduplicationSponsorRecordsFormWizard(UndoDeduplicationRecordsFormWiza
 # Guests
 class SelectAndReviewGuestRecordsFormWizard(SelectAndViewRecordsFormWizard):
     model = MvPerson
-    group_type = [GroupType.DEV]
+    group_type = list(FIX_GUEST_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
     record_type = "guest"
 
     def __init__(self, **kwargs):
@@ -3051,7 +3057,7 @@ class SelectAndReviewGuestRecordsFormWizard(SelectAndViewRecordsFormWizard):
 
 class UndoDeduplicationGuestRecordsFormWizard(UndoDeduplicationRecordsFormWizard):
     model = MvPerson
-    group_type = [GroupType.DEV]
+    group_type = list(FIX_GUEST_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -3239,11 +3245,7 @@ class UndoDeduplicationGuestRecordsFormWizard(UndoDeduplicationRecordsFormWizard
 class SelectAndReviewAccommodationRecordsFormWizard(SelectAndViewRecordsFormWizard):
     model = MvAccommodation
     record_type = "accommodation"
-
-    def get_group_type(self):
-        if settings.FIX_DUPLICATE_RECORDS_ENABLED:
-            return list(FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
-        return [GroupType.DEV]
+    group_type = list(FIX_DUPLICATE_RECORDS_ALLOWED_GROUP_TYPES)
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)