+++ guests/tests/test_detail_overview.py
@@ -135,7 +135,11 @@ class GuestDetailOverviewViewTests(TestSessionTokenMixin, TestCase):
             )
         )
 
-        self.assertContains(response, "Change", html=True)
+        self.assertContains(
+            response,
+            'Change<span class="govuk-visually-hidden">record details</span>',
+            html=True,
+        )
 
         response = self.client.get(
             reverse(
@@ -144,7 +148,11 @@ class GuestDetailOverviewViewTests(TestSessionTokenMixin, TestCase):
             )
         )
 
-        self.assertNotContains(response, "Change", html=True)
+        self.assertNotContains(
+            response,
+            'Change<span class="govuk-visually-hidden">record details</span>',
+            html=True,
+        )
 
     def test_duplicate_label_renders_for_duplicate_guests_only(
         self,